requirements.txt
django==5.0
torch==2.0.1
sentence_transformers==2.2.2
langchain==0.0.266
chromadb==0.4.5
InstructorEmbedding==1.0.1
google-cloud-vision
google-cloud-storage
google-cloud-aiplatform
poppler-utils